Revert "Return long from compiler"
This reverts commit 247d65a1a2d58e2dd0c2c5799006e36c8daaecf0.

+++ include/wisey/Compiler.hpp
@@ -46,7 +46,7 @@ namespace wisey {
     /**
      * Run compiled code
      */
-    long run(int argc, char** argv);
+    int run(int argc, char** argv);
     
   private:
