+++ src/IRGenerationContext.cpp
@@ -88,7 +88,7 @@ IRGenerationContext::~IRGenerationContext() {
   mBindings.clear();
 }
 
-long IRGenerationContext::runCode(int argc, char** argv) {
+int IRGenerationContext::runCode(int argc, char** argv) {
   ExecutionEngine* executionEngine = EngineBuilder(move(mModuleOwner)).create();
   vector<GenericValue> arguments(2);
   arguments[0].IntVal = APInt(32, argc);
@@ -106,7 +106,7 @@ long IRGenerationContext::runCode(int argc, char** argv) {
   Log::i("Result: " + to_string(result));
   delete executionEngine;
   
-  return result;
+  return (int) result;
 }
 
 Module* IRGenerationContext::getModule() {
